feat(vip-616): configurator pattern to fit LZ payload + fix ACM bootstrap
- Cross-chain payload for VIP-616 + VIP-617 combined exceeds the LZ V1
  RelayerV2 ~10 KB ceiling; encode all ACM grants and market wiring inside
  per-chain DeviationSentinelConfigurator so each chain needs only ~6
  cross-chain commands instead of 100+ ACM grants
- Merge vip-617 scope into vip-616 — configurator runs both bootstraps
  atomically in a single execute() call, no second VIP needed
- Fix ACM bootstrap: VIP must grant DEFAULT_ADMIN_ROLE (not per-sig perms);
  Venus ACM.giveCallPermission wraps OZ grantRole which requires that role,
  per-sig grants caused execute() to revert on the first self-grant call
- Add full fork simulation suite for Ethereum, Arbitrum One, and Base:
  pre/post ownership, ACM permissions, market wiring, and end-to-end
  SentinelOracle / checkPriceDeviation assertions
